requirements.txt
contrastive==1.2.0
matplotlib==3.7.5
mlflow==2.13.0
neural==0.1.0
numpy==1.24.3
pandas==2.0.3
pyro_ppl==1.6.0
scipy==1.15.3
seaborn==0.13.2
torch==1.8.0
tqdm==4.66.4
